--- a/src/microtime.js
+++ b/src/microtime.js
@@ -1,6 +1,7 @@
 'use strict'
 
-const time = Date.now() * 1e3
+const time = Date.now()
 const start = process.hrtime.bigint()
 
-module.exports.now = () => time + Number(process.hrtime.bigint() - start) * 1e-3
+// Return high-precision timestamp in milliseconds
+module.exports.now = () => time + Number(process.hrtime.bigint() - start) / 1e6