Sorry for that, but perf matters. Holding this until add a benchmark <picture data-single-emoji=":cowboy-cool-cry-mild-panic:" title=":cowboy-cool-cry-mild-panic:"><img class="emoji" width="20" height="auto" src="https://emoji.slack-edge.com/T0CAQ00TU/cowboy-cool-cry-mild-panic/98654482c63ca15e.png" alt=":cowboy-cool-cry-mild-panic:" align="absmiddle"></picture> 

**before**

```
=== Async RateLimiter Benchmark ===
Iterations: 10000
Concurrency: 100
Rate limit: 100 requests per 60 seconds
IP distribution: 200 IPs (10 hot IPs receiving 20% of traffic)
Redis: localhost:6379
-----------------------------------
Warming up with 1000 requests...
Resetting Redis before benchmark...
Running 10000 iterations...

=== Benchmark Results ===
Total requests: 10000
Successful requests: 10000
Rate limited requests: 0 (0.00%)
Total time: 302.09ms
Requests per second: 33103.24

Latency:
  Average: 1.69ms
  p50: 1.44ms
  p95: 3.93ms
  p99: 6.88ms
```

**after **

```
=== Async RateLimiter Benchmark ===
Iterations: 100000
Concurrency: 100
Rate limit: 100 requests per 60 seconds
IP distribution: 200 IPs (10 hot IPs receiving 20% of traffic)
Redis: localhost:6379
-----------------------------------
Warming up with 1000 requests...
Resetting Redis before benchmark...
Running 100000 iterations...

=== Benchmark Results ===
Total requests: 100000
Successful requests: 97534
Rate limited requests: 2466 (2.47%)
Total time: 1231.12ms
Requests per second: 81226.99

Latency:
  Average: 0.63ms
  p50: 0.61ms
  p95: 0.88ms
  p99: 1.13ms
```