@dy0gu 2025-04-08
Hi there, any plans to merge this? I was thinking about writing my own rate limiter with inlined Lua, much like you just did here. Your solution seems as simple as I was looking to make mine so I would love to have this in a released version!