@Kikobeats 2025-04-08
Thanks for remembering me to ship this @dy0gu 🙂

I added a benchmark, and it seems the library around ~50% faster with the lua implementation.